# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: |
    REST API для управления банковскими картами, пользователями, переводами и аудитом.
    
    ## Основные возможности:
    - Управление пользователями (регистрация, аутентификация, роли)
    - Управление банковскими картами (создание, блокировка, пополнение)
    - Переводы между картами пользователя
    - Система уведомлений для администраторов
    - Аудит действий пользователей
    - Экспорт данных в CSV и PDF форматах
    
    ## Роли пользователей:
    - **ADMIN**: Полный доступ ко всем операциям
    - **USER**: Доступ только к своим картам и переводам
    
    ## Безопасность:
    - JWT токены для аутентификации
    - Шифрование номеров карт в базе данных
    - Валидация данных на уровне сервисов
    - Глобальная обработка ошибок
  version: 1.0.0
  contact:
    name: Bank Cards API Support
    email: support@bankcards.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://api.bankcards.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /admin/export/cards/csv:
    get:
      tags:
        - Export
      summary: Экспорт карт в CSV
      description: Экспортирует данные о картах в CSV формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу карты
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: search
          in: query
          description: Поиск по номеру карты или email
          schema:
            type: string
        - name: ownerEmail
          in: query
          description: Email владельца карты
          schema:
            type: string
      responses:
        '200':
          description: CSV файл с данными о картах
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
  /admin/export/cards/pdf:
    get:
      tags:
        - Export
      summary: Экспорт карт в PDF
      description: Экспортирует данные о картах в PDF формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу карты
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: search
          in: query
          description: Поиск по номеру карты или email
          schema:
            type: string
        - name: ownerEmail
          in: query
          description: Email владельца карты
          schema:
            type: string
      responses:
        '200':
          description: PDF файл с данными о картах
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
  /admin/export/users/csv:
    get:
      tags:
        - Export
      summary: Экспорт пользователей в CSV
      description: Экспортирует данные о пользователях в CSV формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: role
          in: query
          description: Фильтр по роли пользователя
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: search
          in: query
          description: Поиск по имени или email
          schema:
            type: string
      responses:
        '200':
          description: CSV файл с данными о пользователях
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
  /admin/export/users/pdf:
    get:
      tags:
        - Export
      summary: Экспорт пользователей в PDF
      description: Экспортирует данные о пользователях в PDF формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: role
          in: query
          description: Фильтр по роли пользователя
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: search
          in: query
          description: Поиск по имени или email
          schema:
            type: string
      responses:
        '200':
          description: PDF файл с данными о пользователях
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
  /admin/export/audit/csv:
    get:
      tags:
        - Export
      summary: Экспорт аудит-логов в CSV
      description: Экспортирует аудит-логи в CSV формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: action
          in: query
          description: Фильтр по действию
          schema:
            type: string
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
        - name: userEmail
          in: query
          description: Email пользователя
          schema:
            type: string
      responses:
        '200':
          description: CSV файл с аудит-логами
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
  /admin/export/audit/pdf:
    get:
      tags:
        - Export
      summary: Экспорт аудит-логов в PDF
      description: Экспортирует аудит-логи в PDF формате
      security:
        - BearerAuth: [ADMIN]
      parameters:
        - name: action
          in: query
          description: Фильтр по действию
          schema:
            type: string
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
        - name: userEmail
          in: query
          description: Email пользователя
          schema:
            type: string
      responses:
        '200':
          description: PDF файл с аудит-логами
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Недостаточно прав
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
      description: Роль пользователя в системе
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED
      description: Статус банковской карты
...